refactor(KakaoLogin, Group): 카카오 로그인 memberId 수정

See merge request s12-fintech-finance-sub1/S12P21C208!66

+++ backend/src/main/java/com/patriot/fourlipsclover/group/dto/response/GroupResponse.java
@@ -10,11 +10,11 @@ import java.time.LocalDateTime;
 public class GroupResponse {
 
     private Integer groupId;
-    private Integer memberId;
+    private Long memberId;
     private String name;
     private String description;
     private Boolean isPublic;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    
+
 }
